fix: separate feedback into new file, fix footer icon height, fix menu padding bugs, fix feedback turn off success message
