+++ src/course-home/outline-tab/SectionList.jsx
@@ -1,13 +1,13 @@
 /* eslint-disable no-unused-vars */
-import React, { useState } from 'react';
-import PropTypes from 'prop-types';
+import React, { useState } from "react";
+import PropTypes from "prop-types";
 // import {
 //   injectIntl,
 //   intlShape,
 // } from '@edx/frontend-platform/i18n';
 
-import Section from './Section';
-import { useModel } from '../../generic/model-store';
+import Section from "./Section";
+import { useModel } from "../../generic/model-store";
 
 const MAX_HEIGHT_PERCENT = 80;
 
@@ -16,22 +16,28 @@ function SectionList({
   expandAll,
   relativeHeight,
   useHistory,
-  lesson
+  lesson,
 }) {
-
-
   const [height, setHeight] = useState(window.height);
   const resizeObserver = new ResizeObserver(() => {
     // Get height of #section-list-container
-    const sectionListContainer = document.getElementById('section-list-container');
+    const sectionListContainer = document.getElementById(
+      "section-list-container"
+    );
 
     if (relativeHeight && sectionListContainer) {
-      const vh = Math.max(document.documentElement.clientHeight || 0, window.innerHeight || 0);
+      const vh = Math.max(
+        document.documentElement.clientHeight || 0,
+        window.innerHeight || 0
+      );
 
       const vhHeight = Math.round((vh / 100) * MAX_HEIGHT_PERCENT);
       const sectionListContainerHeight = sectionListContainer.offsetHeight;
 
-      const finalHeight = Math.min(vhHeight, Math.round((sectionListContainerHeight / 100) * MAX_HEIGHT_PERCENT));
+      const finalHeight = Math.min(
+        vhHeight,
+        Math.round((sectionListContainerHeight / 100) * MAX_HEIGHT_PERCENT)
+      );
       setHeight(finalHeight);
     }
   });
@@ -39,33 +45,30 @@ function SectionList({
   resizeObserver.observe(document.body);
 
   const {
-    courseBlocks: {
-      courses,
-      sections,
-    },
-  } = useModel('outline', courseId);
+    courseBlocks: { courses, sections },
+  } = useModel("outline", courseId);
 
-const style = {
+  const style = {
     maxHeight: `${height}px`,
-    overflow : 'auto',
+    overflow: "auto",
   };
-
 
   const rootCourseId = courses && Object.keys(courses)[0];
 
   return (
-    <ol id="courseHome-utline" className="list-unstyled" style={style} >
+    <ol id="courseHome-utline" className="list-unstyled" style={style}>
       {courses[rootCourseId].sectionIds.map((sectionId) => {
-
-        return <Section
-        key={sectionId}
-        courseId={courseId}
-        defaultOpen={false}
-        expand={expandAll}
-        section={sections[sectionId]}
-        useHistory={useHistory}
-        lesson = {lesson}
-      />
+        return (
+          <Section
+            key={sectionId}
+            courseId={courseId}
+            defaultOpen={false}
+            expand={expandAll}
+            section={sections[sectionId]}
+            useHistory={useHistory}
+            lesson={lesson}
+          />
+        );
       })}
     </ol>
   );
@@ -76,15 +79,15 @@ SectionList.propTypes = {
   expandAll: PropTypes.bool,
   relativeHeight: PropTypes.bool,
   useHistory: PropTypes.bool,
-  lesson : PropTypes.bool
+  lesson: PropTypes.bool,
 };
 
 SectionList.defaultProps = {
-  courseId: '',
+  courseId: "",
   expandAll: false,
   relativeHeight: false,
   useHistory: false,
-  lesson : false
+  lesson: false,
 };
 
 export default SectionList;
