+++ src/courseware/CoursewareContainer.jsx
@@ -1,9 +1,9 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import { history } from '@edx/frontend-platform';
-import { createSelector } from '@reduxjs/toolkit';
-import { defaultMemoize as memoize } from 'reselect';
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import { connect } from "react-redux";
+import { history } from "@edx/frontend-platform";
+import { createSelector } from "@reduxjs/toolkit";
+import { defaultMemoize as memoize } from "reselect";
 
 import {
   checkBlockCompletion,
@@ -12,122 +12,153 @@ import {
   getResumeBlock,
   getSequenceForUnitDeprecated,
   saveSequencePosition,
-} from './data';
-import { TabPage } from '../tab-page';
+} from "./data";
+import { TabPage } from "../tab-page";
 
-import Course from './course';
-import { handleNextSectionCelebration } from './course/celebration';
-
-// Look at where this is called in componentDidUpdate for more info about its usage
-const checkResumeRedirect = memoize((courseStatus, courseId, sequenceId, firstSequenceId) => {
-  if (courseStatus === 'loaded' && !sequenceId) {
-    // Note that getResumeBlock is just an API call, not a redux thunk.
-    getResumeBlock(courseId).then((data) => {
-      // This is a replace because we don't want this change saved in the browser's history.
-      if (data.sectionId && data.unitId) {
-        history.replace(`/course/${courseId}/${data.sectionId}/${data.unitId}`);
-      } else if (firstSequenceId) {
-        history.replace(`/course/${courseId}/${firstSequenceId}`);
-      }
-    });
-  }
-});
+import Course from "./course";
+import { handleNextSectionCelebration } from "./course/celebration";
 
 // Look at where this is called in componentDidUpdate for more info about its usage
-const checkSectionUnitToUnitRedirect = memoize((courseStatus, courseId, sequenceStatus, section, unitId) => {
-  if (courseStatus === 'loaded' && sequenceStatus === 'failed' && section && unitId) {
-    history.replace(`/course/${courseId}/${unitId}`);
-  }
-});
-
-// Look at where this is called in componentDidUpdate for more info about its usage
-const checkSectionToSequenceRedirect = memoize((courseStatus, courseId, sequenceStatus, section, unitId) => {
-  if (courseStatus === 'loaded' && sequenceStatus === 'failed' && section && !unitId) {
-    // If the section is non-empty, redirect to its first sequence.
-    if (section.sequenceIds && section.sequenceIds[0]) {
-      history.replace(`/course/${courseId}/${section.sequenceIds[0]}`);
-    // Otherwise, just go to the course root, letting the resume redirect take care of things.
-    } else {
-      history.replace(`/course/${courseId}`);
+const checkResumeRedirect = memoize(
+  (courseStatus, courseId, sequenceId, firstSequenceId) => {
+    if (courseStatus === "loaded" && !sequenceId) {
+      // Note that getResumeBlock is just an API call, not a redux thunk.
+      getResumeBlock(courseId).then((data) => {
+        // This is a replace because we don't want this change saved in the browser's history.
+        if (data.sectionId && data.unitId) {
+          history.replace(
+            `/course/${courseId}/${data.sectionId}/${data.unitId}`
+          );
+        } else if (firstSequenceId) {
+          history.replace(`/course/${courseId}/${firstSequenceId}`);
+        }
+      });
     }
   }
-});
+);
 
 // Look at where this is called in componentDidUpdate for more info about its usage
-const checkUnitToSequenceUnitRedirect = memoize((
-  courseStatus, courseId, sequenceStatus, sequenceMightBeUnit, sequenceId, section, routeUnitId,
-) => {
-  if (courseStatus === 'loaded' && sequenceStatus === 'failed' && !section && !routeUnitId) {
-    if (sequenceMightBeUnit) {
-      // If the sequence failed to load as a sequence, but it is marked as a possible unit, then we need to look up the
-      // correct parent sequence for it, and redirect there.
-      const unitId = sequenceId; // just for clarity during the rest of this method
-      getSequenceForUnitDeprecated(courseId, unitId).then(
-        parentId => {
-          if (parentId) {
-            history.replace(`/course/${courseId}/${parentId}/${unitId}`);
-          } else {
+const checkSectionUnitToUnitRedirect = memoize(
+  (courseStatus, courseId, sequenceStatus, section, unitId) => {
+    if (
+      courseStatus === "loaded" &&
+      sequenceStatus === "failed" &&
+      section &&
+      unitId
+    ) {
+      history.replace(`/course/${courseId}/${unitId}`);
+    }
+  }
+);
+
+// Look at where this is called in componentDidUpdate for more info about its usage
+const checkSectionToSequenceRedirect = memoize(
+  (courseStatus, courseId, sequenceStatus, section, unitId) => {
+    if (
+      courseStatus === "loaded" &&
+      sequenceStatus === "failed" &&
+      section &&
+      !unitId
+    ) {
+      // If the section is non-empty, redirect to its first sequence.
+      if (section.sequenceIds && section.sequenceIds[0]) {
+        history.replace(`/course/${courseId}/${section.sequenceIds[0]}`);
+        // Otherwise, just go to the course root, letting the resume redirect take care of things.
+      } else {
+        history.replace(`/course/${courseId}`);
+      }
+    }
+  }
+);
+
+// Look at where this is called in componentDidUpdate for more info about its usage
+const checkUnitToSequenceUnitRedirect = memoize(
+  (
+    courseStatus,
+    courseId,
+    sequenceStatus,
+    sequenceMightBeUnit,
+    sequenceId,
+    section,
+    routeUnitId
+  ) => {
+    if (
+      courseStatus === "loaded" &&
+      sequenceStatus === "failed" &&
+      !section &&
+      !routeUnitId
+    ) {
+      if (sequenceMightBeUnit) {
+        // If the sequence failed to load as a sequence, but it is marked as a possible unit, then we need to look up the
+        // correct parent sequence for it, and redirect there.
+        const unitId = sequenceId; // just for clarity during the rest of this method
+        getSequenceForUnitDeprecated(courseId, unitId).then(
+          (parentId) => {
+            if (parentId) {
+              history.replace(`/course/${courseId}/${parentId}/${unitId}`);
+            } else {
+              history.replace(`/course/${courseId}`);
+            }
+          },
+          () => {
+            // error case
             history.replace(`/course/${courseId}`);
           }
-        },
-        () => { // error case
-          history.replace(`/course/${courseId}`);
-        },
-      );
-    } else {
-      // Invalid sequence that isn't a unit either. Redirect up to main course.
-      history.replace(`/course/${courseId}`);
+        );
+      } else {
+        // Invalid sequence that isn't a unit either. Redirect up to main course.
+        history.replace(`/course/${courseId}`);
+      }
     }
   }
-});
+);
 
 // Look at where this is called in componentDidUpdate for more info about its usage
-const checkSequenceToSequenceUnitRedirect = memoize((courseId, sequenceStatus, sequence, unitId) => {
-  if (sequenceStatus === 'loaded' && sequence.id && !unitId) {
-    if (sequence.unitIds !== undefined && sequence.unitIds.length > 0) {
-      const nextUnitId = sequence.unitIds[sequence.activeUnitIndex];
-      // This is a replace because we don't want this change saved in the browser's history.
-      history.replace(`/course/${courseId}/${sequence.id}/${nextUnitId}`);
+const checkSequenceToSequenceUnitRedirect = memoize(
+  (courseId, sequenceStatus, sequence, unitId) => {
+    if (sequenceStatus === "loaded" && sequence.id && !unitId) {
+      if (sequence.unitIds !== undefined && sequence.unitIds.length > 0) {
+        const nextUnitId = sequence.unitIds[sequence.activeUnitIndex];
+        // This is a replace because we don't want this change saved in the browser's history.
+        history.replace(`/course/${courseId}/${sequence.id}/${nextUnitId}`);
+      }
     }
   }
-});
+);
 
 // Look at where this is called in componentDidUpdate for more info about its usage
-const checkSequenceUnitMarkerToSequenceUnitRedirect = memoize((courseId, sequenceStatus, sequence, unitId) => {
-  if (sequenceStatus !== 'loaded' || !sequence.id) {
-    return;
-  }
-
-  const hasUnits = sequence.unitIds?.length > 0;
-
-  if (unitId === 'first') {
-    if (hasUnits) {
-      const firstUnitId = sequence.unitIds[0];
-      history.replace(`/course/${courseId}/${sequence.id}/${firstUnitId}`);
-    } else {
-      // No units... go to general sequence page
-      history.replace(`/course/${courseId}/${sequence.id}`);
+const checkSequenceUnitMarkerToSequenceUnitRedirect = memoize(
+  (courseId, sequenceStatus, sequence, unitId) => {
+    if (sequenceStatus !== "loaded" || !sequence.id) {
+      return;
     }
-  } else if (unitId === 'last') {
-    if (hasUnits) {
-      const lastUnitId = sequence.unitIds[sequence.unitIds.length - 1];
-      history.replace(`/course/${courseId}/${sequence.id}/${lastUnitId}`);
-    } else {
-      // No units... go to general sequence page
-      history.replace(`/course/${courseId}/${sequence.id}`);
+
+    const hasUnits = sequence.unitIds?.length > 0;
+
+    if (unitId === "first") {
+      if (hasUnits) {
+        const firstUnitId = sequence.unitIds[0];
+        history.replace(`/course/${courseId}/${sequence.id}/${firstUnitId}`);
+      } else {
+        // No units... go to general sequence page
+        history.replace(`/course/${courseId}/${sequence.id}`);
+      }
+    } else if (unitId === "last") {
+      if (hasUnits) {
+        const lastUnitId = sequence.unitIds[sequence.unitIds.length - 1];
+        history.replace(`/course/${courseId}/${sequence.id}/${lastUnitId}`);
+      } else {
+        // No units... go to general sequence page
+        history.replace(`/course/${courseId}/${sequence.id}`);
+      }
     }
   }
-});
+);
 
 class CoursewareContainer extends Component {
   checkSaveSequencePosition = memoize((unitId) => {
-    const {
-      courseId,
-      sequenceId,
-      sequenceStatus,
-      sequence,
-    } = this.props;
-    if (sequenceStatus === 'loaded' && sequence.saveUnitPosition && unitId) {
+    const { courseId, sequenceId, sequenceStatus, sequence } = this.props;
+    if (sequenceStatus === "loaded" && sequence.saveUnitPosition && unitId) {
       const activeUnitIndex = sequence.unitIds.indexOf(unitId);
       this.props.saveSequencePosition(courseId, sequenceId, activeUnitIndex);
     }
@@ -146,10 +177,7 @@ class CoursewareContainer extends Component {
   componentDidMount() {
     const {
       match: {
-        params: {
-          courseId: routeCourseId,
-          sequenceId: routeSequenceId,
-        },
+        params: { courseId: routeCourseId, sequenceId: routeSequenceId },
       },
     } = this.props;
     // Load data whenever the course or sequence ID changes.
@@ -184,7 +212,10 @@ class CoursewareContainer extends Component {
     this.checkSaveSequencePosition(routeUnitId);
 
     // Coerce the route ids into null here because they can be undefined, but the redux ids would be null instead.
-    if (courseId !== (routeCourseId || null) || sequenceId !== (routeSequenceId || null)) {
+    if (
+      courseId !== (routeCourseId || null) ||
+      sequenceId !== (routeSequenceId || null)
+    ) {
       // The non-route ids are pulled from redux state - they are changed at the same time as the status variables.
       // But the route ids are pulled directly from the route. So if the route changes, and we start a fetch above,
       // there's a race condition where the route ids are for one course, but the status and the other ids are for a
@@ -215,73 +246,98 @@ class CoursewareContainer extends Component {
     // otherwise, we could get stuck in a redirect loop, since a sequence that failed to load
     // would endlessly redirect to itself through `checkSectionUnitToUnitRedirect`
     // and `checkUnitToSequenceUnitRedirect`.
-    checkSectionUnitToUnitRedirect(courseStatus, courseId, sequenceStatus, sectionViaSequenceId, routeUnitId);
+    checkSectionUnitToUnitRedirect(
+      courseStatus,
+      courseId,
+      sequenceStatus,
+      sectionViaSequenceId,
+      routeUnitId
+    );
 
     // Check section to sequence redirect:
     //    /course/:courseId/:sectionId         -> /course/:courseId/:sequenceId
     // by redirecting to the first sequence within the section.
-    checkSectionToSequenceRedirect(courseStatus, courseId, sequenceStatus, sectionViaSequenceId, routeUnitId);
+    checkSectionToSequenceRedirect(
+      courseStatus,
+      courseId,
+      sequenceStatus,
+      sectionViaSequenceId,
+      routeUnitId
+    );
 
     // Check unit to sequence-unit redirect:
     //    /course/:courseId/:unitId -> /course/:courseId/:sequenceId/:unitId
     // by filling in the ID of the parent sequence of :unitId.
     checkUnitToSequenceUnitRedirect(
-      courseStatus, courseId, sequenceStatus, sequenceMightBeUnit, sequenceId, sectionViaSequenceId, routeUnitId,
+      courseStatus,
+      courseId,
+      sequenceStatus,
+      sequenceMightBeUnit,
+      sequenceId,
+      sectionViaSequenceId,
+      routeUnitId
     );
 
     // Check sequence to sequence-unit redirect:
     //    /course/:courseId/:sequenceId -> /course/:courseId/:sequenceId/:unitId
     // by filling in the ID the most-recently-active unit in the sequence, OR
     // the ID of the first unit the sequence if none is active.
-    checkSequenceToSequenceUnitRedirect(courseId, sequenceStatus, sequence, routeUnitId);
+    checkSequenceToSequenceUnitRedirect(
+      courseId,
+      sequenceStatus,
+      sequence,
+      routeUnitId
+    );
 
     // Check sequence-unit marker to sequence-unit redirect:
     //    /course/:courseId/:sequenceId/first -> /course/:courseId/:sequenceId/:unitId
     //    /course/:courseId/:sequenceId/last -> /course/:courseId/:sequenceId/:unitId
     // by filling in the ID the first or last unit in the sequence.
     // "Sequence unit marker" is an invented term used only in this component.
-    checkSequenceUnitMarkerToSequenceUnitRedirect(courseId, sequenceStatus, sequence, routeUnitId);
+    checkSequenceUnitMarkerToSequenceUnitRedirect(
+      courseId,
+      sequenceStatus,
+      sequence,
+      routeUnitId
+    );
   }
 
   handleUnitNavigationClick = (nextUnitId) => {
     const {
-      courseId, sequenceId,
+      courseId,
+      sequenceId,
       match: {
-        params: {
-          unitId: routeUnitId,
-        },
+        params: { unitId: routeUnitId },
       },
     } = this.props;
 
     this.props.checkBlockCompletion(courseId, sequenceId, routeUnitId);
     history.push(`/course/${courseId}/${sequenceId}/${nextUnitId}`);
-  }
+  };
 
   handleNextSequenceClick = () => {
-    const {
-      course,
-      courseId,
-      nextSequence,
-      sequence,
-      sequenceId,
-    } = this.props;
+    const { course, courseId, nextSequence, sequence, sequenceId } = this.props;
 
     if (nextSequence !== null) {
       history.push(`/course/${courseId}/${nextSequence.id}/first`);
 
-      const celebrateFirstSection = course && course.celebrations && course.celebrations.firstSection;
-      if (celebrateFirstSection && sequence.sectionId !== nextSequence.sectionId) {
+      const celebrateFirstSection =
+        course && course.celebrations && course.celebrations.firstSection;
+      if (
+        celebrateFirstSection &&
+        sequence.sectionId !== nextSequence.sectionId
+      ) {
         handleNextSectionCelebration(sequenceId, nextSequence.id);
       }
     }
-  }
+  };
 
   handlePreviousSequenceClick = () => {
     const { previousSequence, courseId } = this.props;
     if (previousSequence !== null) {
       history.push(`/course/${courseId}/${previousSequence.id}/last`);
     }
-  }
+  };
 
   render() {
     const {
@@ -289,9 +345,7 @@ class CoursewareContainer extends Component {
       courseId,
       sequenceId,
       match: {
-        params: {
-          unitId: routeUnitId,
-        },
+        params: { unitId: routeUnitId },
       },
     } = this.props;
 
@@ -303,7 +357,7 @@ class CoursewareContainer extends Component {
         courseStatus={courseStatus}
         metadataModel="coursewareMeta"
       >
-          <Course
+        <Course
           courseId={courseId}
           sequenceId={sequenceId}
           unitId={routeUnitId}
@@ -344,8 +398,9 @@ CoursewareContainer.propTypes = {
   courseId: PropTypes.string,
   sequenceId: PropTypes.string,
   firstSequenceId: PropTypes.string,
-  courseStatus: PropTypes.oneOf(['loaded', 'loading', 'failed', 'denied']).isRequired,
-  sequenceStatus: PropTypes.oneOf(['loaded', 'loading', 'failed']).isRequired,
+  courseStatus: PropTypes.oneOf(["loaded", "loading", "failed", "denied"])
+    .isRequired,
+  sequenceStatus: PropTypes.oneOf(["loaded", "loading", "failed"]).isRequired,
   sequenceMightBeUnit: PropTypes.bool.isRequired,
   nextSequence: sequenceShape,
   previousSequence: sequenceShape,
@@ -372,13 +427,15 @@ CoursewareContainer.defaultProps = {
 const currentCourseSelector = createSelector(
   (state) => state.models.coursewareMeta || {},
   (state) => state.courseware.courseId,
-  (coursesById, courseId) => (coursesById[courseId] ? coursesById[courseId] : null),
+  (coursesById, courseId) =>
+    coursesById[courseId] ? coursesById[courseId] : null
 );
 
 const currentSequenceSelector = createSelector(
   (state) => state.models.sequences || {},
   (state) => state.courseware.sequenceId,
-  (sequencesById, sequenceId) => (sequencesById[sequenceId] ? sequencesById[sequenceId] : null),
+  (sequencesById, sequenceId) =>
+    sequencesById[sequenceId] ? sequencesById[sequenceId] : null
 );
 
 const sequenceIdsSelector = createSelector(
@@ -386,12 +443,14 @@ const sequenceIdsSelector = createSelector(
   currentCourseSelector,
   (state) => state.models.sections,
   (courseStatus, course, sectionsById) => {
-    if (courseStatus !== 'loaded') {
+    if (courseStatus !== "loaded") {
       return [];
     }
     const { sectionIds = [] } = course;
-    return sectionIds.flatMap(sectionId => sectionsById[sectionId].sequenceIds);
-  },
+    return sectionIds.flatMap(
+      (sectionId) => sectionsById[sectionId].sequenceIds
+    );
+  }
 );
 
 const previousSequenceSelector = createSelector(
@@ -403,9 +462,12 @@ const previousSequenceSelector = createSelector(
       return null;
     }
     const sequenceIndex = sequenceIds.indexOf(sequenceId);
-    const previousSequenceId = sequenceIndex > 0 ? sequenceIds[sequenceIndex - 1] : null;
-    return previousSequenceId !== null ? sequencesById[previousSequenceId] : null;
-  },
+    const previousSequenceId =
+      sequenceIndex > 0 ? sequenceIds[sequenceIndex - 1] : null;
+    return previousSequenceId !== null
+      ? sequencesById[previousSequenceId]
+      : null;
+  }
 );
 
 const nextSequenceSelector = createSelector(
@@ -417,9 +479,12 @@ const nextSequenceSelector = createSelector(
       return null;
     }
     const sequenceIndex = sequenceIds.indexOf(sequenceId);
-    const nextSequenceId = sequenceIndex < sequenceIds.length - 1 ? sequenceIds[sequenceIndex + 1] : null;
+    const nextSequenceId =
+      sequenceIndex < sequenceIds.length - 1
+        ? sequenceIds[sequenceIndex + 1]
+        : null;
     return nextSequenceId !== null ? sequencesById[nextSequenceId] : null;
-  },
+  }
 );
 
 const firstSequenceIdSelector = createSelector(
@@ -427,7 +492,7 @@ const firstSequenceIdSelector = createSelector(
   currentCourseSelector,
   (state) => state.models.sections || {},
   (courseStatus, course, sectionsById) => {
-    if (courseStatus !== 'loaded') {
+    if (courseStatus !== "loaded") {
       return null;
     }
     const { sectionIds = [] } = course;
@@ -437,13 +502,14 @@ const firstSequenceIdSelector = createSelector(
     }
 
     return sectionsById[sectionIds[0]].sequenceIds[0];
-  },
+  }
 );
 
 const sectionViaSequenceIdSelector = createSelector(
   (state) => state.models.sections || {},
   (state) => state.courseware.sequenceId,
-  (sectionsById, sequenceId) => (sectionsById[sequenceId] ? sectionsById[sequenceId] : null),
+  (sectionsById, sequenceId) =>
+    sectionsById[sequenceId] ? sectionsById[sequenceId] : null
 );
 
 const mapStateToProps = (state) => {
