+++ src/footer/Footer.scss
@@ -23,7 +23,7 @@
         & img {
           display: block;
           width: auto;
-          height: 2.5rem;
+          height: 2rem;
           object-fit: cover;
         }
       }
