+++ src/index.scss
@@ -1737,7 +1737,7 @@ html:root {
 
 .header-title {
   flex-grow: 0;
-  font-family: RobotoSerif;
+  font-family: "Roboto Serif", serif;
   font-size: 2rem;
   font-weight: 500;
   font-stretch: normal;
